Separate code licensed under Apache License into separate file
This is more transparent and also makes copyright update script work

DEVSIX-3611
